Pulling validation out into a repeatable method
